+++ pybamm/models/event.py
@@ -46,6 +46,28 @@ class Event:
         self._expression = expression
         self._event_type = event_type
 
+    @classmethod
+    def _from_json(cls, snippet: dict):
+        """
+        Reconstructs an Event instance during deserialisation of a JSON file.
+
+        Parameters
+        ----------
+        snippet: dict
+            Contains the information needed to reconstruct a specific instance.
+            Should contain "name", "expression" and "event_type".
+        """
+
+        instance = cls.__new__(cls)
+
+        instance.__init__(
+            snippet["name"],
+            snippet["expression"],
+            event_type=EventType(snippet["event_type"][1]),
+        )
+
+        return instance
+
     def evaluate(self, t=None, y=None, y_dot=None, inputs=None):
         """
         Acts as a drop-in replacement for :func:`pybamm.Symbol.evaluate`
@@ -66,3 +88,21 @@ class Event:
     @property
     def event_type(self):
         return self._event_type
+
+    def to_json(self):
+        """
+        Method to serialise an Event object into JSON.
+
+        The expression is written out seperately,
+        See :meth:`pybamm.Serialise._SymbolEncoder.default()`
+        """
+
+        # event_type contains string name, for JSON readability,
+        # and value for deserialisation.
+
+        json_dict = {
+            "name": self._name,
+            "event_type": [str(self._event_type), self._event_type.value],
+        }
+
+        return json_dict