+++ scripts/update_version.py
@@ -30,6 +30,16 @@ def update_version():
         file.seek(0)
         file.write(replace_version)
 
+    # pyproject.toml
+    with open(os.path.join(pybamm.root_dir(), "pyproject.toml"), "r+") as file:
+        output = file.read()
+        replace_version = re.sub(
+            '(?<=version = ")(.+)(?=")', release_version, output
+        )
+        file.truncate(0)
+        file.seek(0)
+        file.write(replace_version)
+
     # CITATION.cff
     with open(os.path.join(pybamm.root_dir(), "CITATION.cff"), "r+") as file:
         output = file.read()
@@ -37,26 +47,6 @@ def update_version():
         file.truncate(0)
         file.seek(0)
         file.write(replace_version)
-
-    # docs/source/_static/versions.json for readthedocs build
-    if "rc" not in release_version:
-        with open(
-            os.path.join(pybamm.root_dir(), "docs", "_static", "versions.json"),
-            "r+",
-        ) as file:
-            output = file.read()
-            json_data = json.loads(output)
-            json_data.insert(
-                2,
-                {
-                    "name": f"v{release_version}",
-                    "version": f"{release_version}",
-                    "url": f"https://docs.pybamm.org/en/v{release_version}/",
-                },
-            )
-            file.truncate(0)
-            file.seek(0)
-            file.write(json.dumps(json_data, indent=4))
 
     # vcpkg.json
     with open(os.path.join(pybamm.root_dir(), "vcpkg.json"), "r+") as file:
@@ -68,7 +58,7 @@ def update_version():
         file.write(replace_version)
 
     # Get latest commit id from pybamm-team/sundials-vcpkg-registry
-    cmd = "git ls-remote https://github.com/pybamm-team/sundials-vcpkg-registry | grep refs/heads/main | cut -f 1 | tr -d '\n'"  # noqa: E501
+    cmd = "git ls-remote https://github.com/pybamm-team/sundials-vcpkg-registry | grep refs/heads/main | cut -f 1 | tr -d '\n'"
     latest_commit_id = os.popen(cmd).read()
 
     # vcpkg-configuration.json
@@ -83,7 +73,7 @@ def update_version():
         file.write(replace_commit_id)
 
     changelog_line1 = "# [Unreleased](https://github.com/pybamm-team/PyBaMM/)\n"
-    changelog_line2 = f"# [v{release_version}](https://github.com/pybamm-team/PyBaMM/tree/v{release_version}) - {last_day_of_month}\n\n"  # noqa: E501
+    changelog_line2 = f"# [v{release_version}](https://github.com/pybamm-team/PyBaMM/tree/v{release_version}) - {last_day_of_month}\n\n"
 
     # CHANGELOG.md
     with open(os.path.join(pybamm.root_dir(), "CHANGELOG.md"), "r+") as file: